% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_DiffFracSeq_model.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Creates the input data for the stan model}
\usage{
create_stan_data(count_data)
}
\arguments{
\item{count_data}{A data frame of transcript counts with columns: total,
sup, pel, gene_name, rep, and condition.}
}
\value{
A list with values NRNA, NREP, NCON, tot_obs, sup_obs, and pel_obs.
}
\description{
Creates the input list object from a tibble of transcripts counts
required for stan to train the DiffFracSeq model
}
