% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_two_fraction_counts}
\alias{simulated_two_fraction_counts}
\title{Simulated fraction count data}
\format{
\subsection{\code{simulated_two_fraction_counts}}{

A data frame with 900 rows and 6 columns:
\describe{
\item{gene_name}{Gene name}
\item{rep}{Replicate number}
\item{condition}{name of Condtion}
\item{total}{RNA trancript counts from the total sample}
\item{sup}{RNA trancript counts from the sup sample}
\item{pel}{RNA trancript counts from the pel sample}
...
}
}
}
\source{
Simulated RNA-Seq data
}
\usage{
simulated_two_fraction_counts
}
\description{
Two fraction count data to test the DiffFracSeq model
}
\keyword{datasets}
