% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_DiffFracSeq_model.R
\name{test_frac_across_condition}
\alias{test_frac_across_condition}
\title{Compares fraction counts between conditions}
\usage{
test_frac_across_condition(condition, ratio)
}
\arguments{
\item{condition}{A vector of equal size to ratio stating which condition that ratio was taken from.}

\item{ratio}{A vector of equal size to condition stating the ratio of counts between the two fractions}
}
\value{
A tibble of pvalues comparing rations between all pairs of conditions with three columns:
condition_A, condition_B, and p.value.
}
\description{
Determines if the ratio of counts between two fractions changes significantlu
across all permutations of conditioms
}
