% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_DiffFracSeq_model.R
\name{train_DiffFracSeq_model}
\alias{train_DiffFracSeq_model}
\title{Trains DiffFracSeq model using the input data}
\usage{
train_DiffFracSeq_model(count_data, nchain = 4, ...)
}
\arguments{
\item{count_data}{A data frame of transcript counts with columns: total,
sup, pel, gene_name, rep, and condition.}

\item{nchain}{the number of chains to run when fittting the model}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
First calls the create_stan_data function to create the stan input list object
then calls the rstan::sampling function to learn the model parameters using
the data set
}
